+++ one/tests/alf/test_cache.py
@@ -21,7 +21,7 @@ class TestsONEParquet(unittest.TestCase):
         'subject': 'mysub',
         'date': datetime.date.fromisoformat('2021-02-28'),
         'number': int('001'),
-        'project': '',
+        'projects': '',
         'task_protocol': '',
         'id': 'mylab/mysub/2021-02-28/001',
     }
@@ -53,10 +53,6 @@ class TestsONEParquet(unittest.TestCase):
         self.assertEqual(apt._get_session_info(self.rel_ses_path), self.ses_info)
         self.assertTrue(
             self.full_ses_path.as_posix().endswith(self.rel_ses_path[:-1]))
-
-    def test_walk_session(self):
-        ses_files = list(apt._iter_datasets(self.full_ses_path))
-        self.assertEqual(ses_files, self.rel_ses_files)
 
     def test_parquet(self):
         # Test data