+++ one/tests/remote/test_remote_aws.py
@@ -8,11 +8,11 @@ from pathlib import Path
 from functools import partial
 
 from one.tests import TEST_DB_1, OFFLINE_ONLY, util
-from one.api import OneAlyx
+from one.webclient import AlyxClient
 try:
     import one.remote.aws as aws
 except ModuleNotFoundError:
-    unittest.skip('boto3 module not installed')
+    raise unittest.SkipTest('boto3 module not installed')
 
 
 @unittest.skipIf(OFFLINE_ONLY, 'online only test')
@@ -69,15 +69,14 @@ class TestAWS(unittest.TestCase):
         cls.tempdir = util.set_up_env()
         with mock.patch('one.params.iopar.getfile', new=partial(util.get_file, cls.tempdir.name)):
             # util.setup_test_params(token=True)
-            cls.one = OneAlyx(
+            cls.alyx = AlyxClient(
                 **TEST_DB_1,
-                cache_dir=cls.tempdir.name,
-                mode='local'
+                cache_dir=cls.tempdir.name
             )
 
     def test_credentials(self):
         """Test for one.remote.aws.get_aws_access_keys function."""
-        cred, bucket_name = aws.get_aws_access_keys(self.one, 'aws_cortexlab')
+        cred, bucket_name = aws.get_aws_access_keys(self.alyx, 'aws_cortexlab')
         expected = {
             'aws_access_key_id': 'ABCDEF',
             'aws_secret_access_key': 'shhh',
@@ -85,3 +84,23 @@ class TestAWS(unittest.TestCase):
         }
         self.assertDictEqual(cred, expected)
         self.assertEqual(bucket_name, 's3_bucket')
+
+
+class TestUtils(unittest.TestCase):
+    """Tests for one.remote.aws utility functions"""
+
+    def test_get_s3_virtual_host(self):
+        """Tests for one.remote.aws.get_s3_virtual_host function"""
+        expected = 'https://my-s3-bucket.s3.eu-east-1.amazonaws.com/'
+        url = aws.get_s3_virtual_host('s3://my-s3-bucket', 'eu-east-1')
+        self.assertEqual(expected, url)
+
+        # NB slight diff in behaviour: with no scheme provided output URL has no trailing slash.
+        url = aws.get_s3_virtual_host('my-s3-bucket/', 'eu-east-1')
+        self.assertEqual(expected.strip('/'), url)
+        expected = 'https://my-s3-bucket.s3.eu-east-1.amazonaws.com/path/to/file'
+        url = aws.get_s3_virtual_host('s3://my-s3-bucket/path/to/file', 'eu-east-1')
+        self.assertEqual(expected, url)
+
+        with self.assertRaises(AssertionError):
+            aws.get_s3_virtual_host('s3://my-s3-bucket/path/to/file', 'wrong-foo-4')